test: allow tests to pass without internet
Currently when running the test without an internet connection there are
two JavaScript test failures

This commit moves the two JavaScript tests to test/internet.

Backport-PR-URL: https://github.com/nodejs/node/pull/17171
PR-URL: https://github.com/nodejs/node/pull/16255
